wd: check if sess sched key alloc failed
Alloc sched key may failed, so use WD_IS_ERR
to check return pointer.

+++ include/wd.h
@@ -76,8 +76,8 @@ extern FILE *flog_fd;
 #define	WD_ENOPROC			68 /* no processed */
 
 #define WD_HANDLE_ERR(h)		((long long)(h))
-#define WD_IS_ERR(h)			((unsigned long long)(h) > \
-					(unsigned long long)(-1000))
+#define WD_IS_ERR(h)			((uintptr_t)(h) > \
+					(uintptr_t)(-1000))
 
 static inline void *WD_ERR_PTR(uintptr_t error)
 {
